Cleaned up code in GameScene.ts and moved required code to Network.ts
